Fixed tmux session naming over ssh
the basename of the PPID of the bash process spawning tmux is fine for local
shell accounts but when the ppid is an sshd process the return from ps is not a
file path.

Added a check to see if its a file path and if not just take the first
alphanumeric string ofr the session name.